refactor: move _binop() util function to be near ir.Value, not ir.Expr
This function is only uses with Values, not general Expr,
so it should be co-located there.

This adds a runtime assert to verify we are using this correctly internally.

It also improves the docstring to not be misleading about ops.Binary,
and to give some more examples of the edge cases.
